refactor: useRecoilCallback를 활용한 상태 업데이트 로직 개선
- useSetRecoilState에서 인자를 반드시 넘겨줘야 하는 문제를 해결하기 위해 useRecoilCallback을 사용

- https://recoiljs.org/docs/api-reference/core/useRecoilCallback/
