+++ src/components/CartController/CartController.tsx
@@ -1,5 +1,5 @@
 import { ChangeEvent } from "react";
-import type { ProductItem } from "../../types/types";
+import type { NewCartItem, ProductItem } from "../../types/types";
 import {
   AddCartButton,
   CartBox,
@@ -7,12 +7,10 @@ import {
   QuantityControlButton,
   QuantityInput,
 } from "./CartController.style";
-import {
-  addCartItemSelector,
-  quantityByProductIdSelector,
-  updateCartItemQuantitySelector,
-} from "../../recoil/cartAtoms";
-import { useRecoilValue, useSetRecoilState } from "recoil";
+import { quantityByProductIdSelector } from "../../recoil/cartAtoms";
+import { useRecoilCallback, useRecoilValue } from "recoil";
+import { cartState } from "../../recoil/cartAtoms";
+import { fetchAddCart, fetchDeleteCart, fetchUpdateCart } from "../../api/api";
 
 interface CartControllerProps {
   product: ProductItem;
@@ -20,11 +18,60 @@ interface CartControllerProps {
 
 function CartController({ product }: CartControllerProps) {
   const quantity = useRecoilValue(quantityByProductIdSelector(product.id));
-  const updateCartItemQuantity = useSetRecoilState(
-    updateCartItemQuantitySelector(product.id)
+
+  const updateCartItemQuantity = useRecoilCallback(
+    ({ snapshot, set }) =>
+      async (productId: number, quantity: number) => {
+        const cartList = await snapshot.getPromise(cartState);
+        if (quantity === 0) {
+          const id = productId as number;
+          if (confirm("정말로 삭제하시겠습니까?")) {
+            const removedCartList = cartList.filter((cart) => cart.id !== id);
+            set(cartState, removedCartList);
+            fetchDeleteCart(id);
+          }
+        } else {
+          const targetIndex = cartList.findIndex(
+            (cartItem) => cartItem.id === productId
+          );
+
+          if (targetIndex !== -1) {
+            const updatedCartList = [...cartList];
+            updatedCartList[targetIndex] = {
+              ...updatedCartList[targetIndex],
+              quantity,
+            };
+            set(cartState, updatedCartList);
+
+            fetchUpdateCart(productId, quantity);
+          }
+        }
+      },
+    []
   );
 
-  const addCartItem = useSetRecoilState(addCartItemSelector(undefined));
+  const addCartItem = useRecoilCallback(
+    ({ snapshot, set }) =>
+      async (product: ProductItem) => {
+        const cartList = await snapshot.getPromise(cartState);
+        const isCartItemExist = cartList.some(
+          (cartItem) => cartItem.id === product.id
+        );
+
+        if (!isCartItemExist) {
+          const newCartItem: NewCartItem = {
+            id: product.id,
+            quantity: 1,
+            checked: true,
+            product,
+          };
+          const updatedCartList = [...cartList, newCartItem];
+          set(cartState, updatedCartList);
+          fetchAddCart(newCartItem.id);
+        }
+      },
+    []
+  );
 
   const handleChangeQuantity = (event: ChangeEvent<HTMLInputElement>) => {
     const quantityInputValue = Number(
@@ -32,7 +79,7 @@ function CartController({ product }: CartControllerProps) {
     );
     const newQuantity = quantityInputValue > 100 ? 100 : quantityInputValue;
 
-    updateCartItemQuantity(newQuantity);
+    updateCartItemQuantity(product.id, newQuantity);
   };
 
   return (
@@ -42,7 +89,7 @@ function CartController({ product }: CartControllerProps) {
           <CartBox>
             <QuantityControlButton
               onClick={() => {
-                updateCartItemQuantity(quantity - 1);
+                updateCartItemQuantity(product.id, quantity - 1);
               }}
             >
               -
@@ -50,7 +97,7 @@ function CartController({ product }: CartControllerProps) {
             <QuantityInput value={quantity} onChange={handleChangeQuantity} />
             <QuantityControlButton
               onClick={() => {
-                updateCartItemQuantity(quantity + 1);
+                updateCartItemQuantity(product.id, quantity + 1);
               }}
             >
               +
