+++ src/components/CartList/CartList.tsx
@@ -1,11 +1,10 @@
-import { useRecoilValue, useSetRecoilState } from "recoil";
+import { useRecoilCallback, useRecoilValue } from "recoil";
 import {
   allCartCheckedSelector,
   cartCountSelector,
   cartState,
   checkedCartCountSelector,
-  removeCartItemsSelector,
-  switchAllCartCheckboxSelector,
+  checkedCartSelector,
 } from "../../recoil/cartAtoms";
 import CartItem from "../CartItem";
 import {
@@ -14,6 +13,8 @@ import {
   CartListWrapper,
   CartsDeleteButton,
 } from "./CartList.style";
+import { ReceivedCartItem } from "../../types/types.ts";
+import { fetchDeleteCart } from "../../api/api.ts";
 
 function CartList() {
   const cartList = useRecoilValue(cartState);
@@ -21,8 +22,41 @@ function CartList() {
   const cartCount = useRecoilValue(cartCountSelector);
   const isAllCartItemChecked = useRecoilValue(allCartCheckedSelector);
 
-  const removeCheckedCartItems = useSetRecoilState(removeCartItemsSelector);
-  const switchAllCheckboxes = useSetRecoilState(switchAllCartCheckboxSelector);
+  const removeCheckedCartItems = useRecoilCallback(
+    ({ snapshot, set }) =>
+      async () => {
+        const cartList = await snapshot.getPromise(cartState);
+        const checkedCartList = await snapshot.getPromise(checkedCartSelector);
+        if (confirm("정말로 삭제하시겠습니까?")) {
+          const targetIds = checkedCartList.map((cartList) => cartList.id);
+          const removedCartList = cartList.filter(
+            (cart) => !targetIds.includes(cart.id)
+          );
+          set(cartState, removedCartList);
+          targetIds.forEach((id) => {
+            fetchDeleteCart(id);
+          });
+        }
+      },
+    []
+  );
+
+  const switchAllCheckboxes = useRecoilCallback(
+    ({ snapshot, set }) =>
+      async () => {
+        const cartList = await snapshot.getPromise(cartState);
+        const isAllCartItemChecked = await snapshot.getPromise(
+          allCartCheckedSelector
+        );
+        const newCartList = cartList.map((cartItem: ReceivedCartItem) => ({
+          ...cartItem,
+          checked: !isAllCartItemChecked,
+        }));
+        set(cartState, newCartList);
+      },
+    []
+  );
+
   return (
     <CartListWrapper>
       {cartList.map((cart) => (
@@ -32,12 +66,12 @@ function CartList() {
         <input
           type="checkbox"
           checked={isAllCartItemChecked}
-          onChange={() => switchAllCheckboxes(undefined)}
+          onChange={() => switchAllCheckboxes()}
         />
-        <CartListCheckCounter onClick={() => switchAllCheckboxes(undefined)}>
+        <CartListCheckCounter onClick={() => switchAllCheckboxes()}>
           전체선택 ({checkedCartListCount}/{cartCount})
         </CartListCheckCounter>
-        <CartsDeleteButton onClick={() => removeCheckedCartItems(undefined)}>
+        <CartsDeleteButton onClick={() => removeCheckedCartItems()}>
           선택삭제
         </CartsDeleteButton>
       </CartListController>
